WIP:Move drawing code into LayoutTrackView subclasses
