+++ java/src/jmri/jmrit/display/layoutEditor/LayoutTurntable.java
@@ -2,25 +2,20 @@ package jmri.jmrit.display.layoutEditor;
 
 import static java.lang.Float.POSITIVE_INFINITY;
 
-import java.awt.BasicStroke;
-import java.awt.Color;
-import java.awt.Graphics2D;
-import java.awt.Stroke;
 import java.awt.event.ActionEvent;
 import java.awt.event.MouseEvent;
-import java.awt.geom.Ellipse2D;
-import java.awt.geom.Line2D;
 import java.awt.geom.Point2D;
 import java.awt.geom.Rectangle2D;
 import java.beans.*;
 import java.text.MessageFormat;
 import java.util.*;
+
 import javax.annotation.CheckForNull;
 import javax.annotation.Nonnull;
 import javax.swing.*;
+
 import jmri.*;
 import jmri.util.MathUtil;
-import org.slf4j.*;
 
 /**
  * A LayoutTurntable is a representation used by LayoutEditor to display a
@@ -1105,7 +1100,7 @@ public class LayoutTurntable extends LayoutTrack {
          *
          * @return true if occupied
          */
-        private boolean isOccupied() {
+        public boolean isOccupied() {
             boolean result = false; // assume not
             if (connect != null) {  // does it have a connection? (yes)
                 LayoutBlock lb = connect.getLayoutBlock();
@@ -1221,200 +1216,6 @@ public class LayoutTurntable extends LayoutTrack {
             }
         }
     }   // class RayTrack
-
-    /**
-     * Draw track decorations.
-     * 
-     * This type of track has none, so this method is empty.
-     */
-    @Override
-    protected void drawDecorations(Graphics2D g2) {}
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void draw1(Graphics2D g2, boolean isMain, boolean isBlock) {
-        float trackWidth = 2.F;
-        float halfTrackWidth = trackWidth / 2.f;
-        double radius = getRadius(), diameter = 2.f * radius;
-
-        if (isBlock && isMain) {
-            double radius2 = Math.max(radius / 4.f, trackWidth * 2);
-            double diameter2 = radius2 * 2.f;
-            Stroke stroke = g2.getStroke();
-            Color color = g2.getColor();
-            // draw turntable circle - default track color, side track width
-            g2.setStroke(new BasicStroke(trackWidth, BasicStroke.CAP_BUTT, BasicStroke.JOIN_ROUND));
-            g2.setColor(layoutEditor.getDefaultTrackColorColor());
-            g2.draw(new Ellipse2D.Double(getCoordsCenter().getX() - radius, getCoordsCenter().getY() - radius, diameter, diameter));
-            g2.draw(new Ellipse2D.Double(getCoordsCenter().getX() - radius2, getCoordsCenter().getY() - radius2, diameter2, diameter2));
-            g2.setStroke(stroke);
-            g2.setColor(color);
-        }
-
-        // draw ray tracks
-        for (int j = 0; j < getNumberRays(); j++) {
-            boolean main = false;
-            Color color = null;
-            TrackSegment ts = getRayConnectOrdered(j);
-            if (ts != null) {
-                main = ts.isMainline();
-            }
-
-            if (isBlock) {
-                if (ts == null) {
-                    g2.setColor(layoutEditor.getDefaultTrackColorColor());
-                } else {
-                    LayoutBlock lb = ts.getLayoutBlock();
-                    if (lb != null) {
-                        color = g2.getColor();
-                        setColorForTrackBlock(g2, lb);
-                    }
-                }
-            }
-
-            Point2D pt2 = getRayCoordsOrdered(j);
-            Point2D delta = MathUtil.normalize(MathUtil.subtract(pt2, getCoordsCenter()), radius);
-            Point2D pt1 = MathUtil.add(getCoordsCenter(), delta);
-            if (main == isMain) {
-                g2.draw(new Line2D.Double(pt1, pt2));
-            }
-            if (isMain && isTurnoutControlled() && (getPosition() == j)) {
-                if (isBlock) {
-                    LayoutBlock lb = getLayoutBlock();
-                    if (lb != null) {
-                        color = (color == null) ? g2.getColor() : color;
-                        setColorForTrackBlock(g2, lb);
-                    } else {
-                        g2.setColor(layoutEditor.getDefaultTrackColorColor());
-                    }
-                }
-                delta = MathUtil.normalize(delta, radius - halfTrackWidth);
-                pt1 = MathUtil.subtract(getCoordsCenter(), delta);
-                g2.draw(new Line2D.Double(pt1, pt2));
-            }
-            if (color != null) {
-                g2.setColor(color); /// restore previous color
-            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void draw2(Graphics2D g2, boolean isMain, float railDisplacement) {
-        float trackWidth = 2.F;
-        float halfTrackWidth = trackWidth / 2.f;
-
-        // draw ray tracks
-        for (int j = 0; j < getNumberRays(); j++) {
-            boolean main = false;
-//            Color c = null;
-            TrackSegment ts = getRayConnectOrdered(j);
-            if (ts != null) {
-                main = ts.isMainline();
-//                LayoutBlock lb = ts.getLayoutBlock();
-//                if (lb != null) {
-//                    c = g2.getColor();
-//                    setColorForTrackBlock(g2, lb);
-//                }
-            }
-            Point2D pt2 = getRayCoordsOrdered(j);
-            Point2D vDelta = MathUtil.normalize(MathUtil.subtract(pt2, getCoordsCenter()), radius);
-            Point2D vDeltaO = MathUtil.normalize(MathUtil.orthogonal(vDelta), railDisplacement);
-            Point2D pt1 = MathUtil.add(getCoordsCenter(), vDelta);
-            Point2D pt1L = MathUtil.subtract(pt1, vDeltaO);
-            Point2D pt1R = MathUtil.add(pt1, vDeltaO);
-            Point2D pt2L = MathUtil.subtract(pt2, vDeltaO);
-            Point2D pt2R = MathUtil.add(pt2, vDeltaO);
-            if (main == isMain) {
-                g2.draw(new Line2D.Double(pt1L, pt2L));
-                g2.draw(new Line2D.Double(pt1R, pt2R));
-            }
-            if (isMain && isTurnoutControlled() && (getPosition() == j)) {
-//                LayoutBlock lb = getLayoutBlock();
-//                if (lb != null) {
-//                    c = g2.getColor();
-//                    setColorForTrackBlock(g2, lb);
-//                } else {
-//                    g2.setColor(layoutEditor.getDefaultTrackColorColor());
-//                }
-                vDelta = MathUtil.normalize(vDelta, radius - halfTrackWidth);
-                pt1 = MathUtil.subtract(getCoordsCenter(), vDelta);
-                pt1L = MathUtil.subtract(pt1, vDeltaO);
-                pt1R = MathUtil.add(pt1, vDeltaO);
-                g2.draw(new Line2D.Double(pt1L, pt2L));
-                g2.draw(new Line2D.Double(pt1R, pt2R));
-            }
-//            if (c != null) {
-//                g2.setColor(c); /// restore previous color
-//            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void highlightUnconnected(Graphics2D g2, HitPointType specificType) {
-        for (int j = 0; j < getNumberRays(); j++) {
-            if (  (specificType == HitPointType.NONE)
-                    || (specificType == (HitPointType.turntableTrackIndexedValue(j)))
-                ) 
-            {
-                if (getRayConnectOrdered(j) == null) {
-                    Point2D pt = getRayCoordsOrdered(j);
-                    g2.fill(trackControlCircleAt(pt));
-                }
-            }
-        }
-    }
-
-    /**
-     * Draw this turntable's controls.
-     *
-     * @param g2 the graphics port to draw to
-     */
-    @Override
-    protected void drawTurnoutControls(Graphics2D g2) {
-        if (isTurnoutControlled()) {
-            // draw control circles at all but current position ray tracks
-            for (int j = 0; j < getNumberRays(); j++) {
-                if (getPosition() != j) {
-                    RayTrack rt = rayTrackList.get(j);
-                    if (!rt.isDisabled() && !(rt.isDisabledWhenOccupied() && rt.isOccupied())) {
-                        Point2D pt = getRayCoordsOrdered(j);
-                        g2.draw(trackControlCircleAt(pt));
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * Draw this turntable's edit controls.
-     *
-     * @param g2 the graphics port to draw to
-     */
-    @Override
-    protected void drawEditControls(Graphics2D g2) {
-        Point2D pt = getCoordsCenter();
-        g2.setColor(layoutEditor.getDefaultTrackColorColor());
-        g2.draw(trackControlCircleAt(pt));
-
-        for (int j = 0; j < getNumberRays(); j++) {
-            pt = getRayCoordsOrdered(j);
-
-            if (getRayConnectOrdered(j) == null) {
-                g2.setColor(Color.red);
-            } else {
-                g2.setColor(Color.green);
-            }
-            g2.draw(layoutEditor.layoutEditorControlRectAt(pt));
-        }
-    }
 
     /**
      * {@inheritDoc}
