+++ java/src/jmri/jmrit/display/layoutEditor/LayoutXOver.java
@@ -1,31 +1,14 @@
 package jmri.jmrit.display.layoutEditor;
 
-import static java.lang.Float.POSITIVE_INFINITY;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import java.awt.Color;
-import java.awt.Graphics2D;
-import java.awt.event.ActionEvent;
-import java.awt.event.MouseEvent;
-import java.awt.geom.GeneralPath;
-import java.awt.geom.Line2D;
 import java.awt.geom.Point2D;
-import java.awt.geom.Rectangle2D;
-import java.text.MessageFormat;
-import java.util.*;
+
 import javax.annotation.*;
-import javax.swing.*;
-import jmri.*;
-import jmri.jmrit.display.layoutEditor.blockRoutingTable.LayoutBlockRouteTableAction;
-import jmri.jmrit.signalling.SignallingGuiTools;
-import jmri.util.MathUtil;
-import org.slf4j.*;
 
 /**
- * A LayoutXOver corresponds to a crossover (connection between parallel tracks) on the layout. 
+ * A LayoutXOver corresponds to a crossover (connection between parallel tracks) on the layout.
  * <p>
  * Three types are supported: double crossover,
- * right-handed single crossover, and left-handed single crossover. 
+ * right-handed single crossover, and left-handed single crossover.
  * <p>
  * A LayoutXOver has four connection points, designated A, B, C, and
  * D. The A-B and C-D routes are a straight segment (continuing
@@ -53,7 +36,7 @@ import org.slf4j.*;
  * is assigned, that block applies to the entire turnout.
  * <p>
  * For drawing purposes, each LayoutXOver carries a center point and
- * displacements for B and C. 
+ * displacements for B and C.
  * For double crossovers, the center point is at the center of the turnout, and
  * the displacement for A = - the displacement for C and the displacement for D
  * = - the displacement for B. The center point and these displacements may be
@@ -86,9 +69,9 @@ abstract public class LayoutXOver extends LayoutTurnout {
     public LayoutXOver(@Nonnull String id, TurnoutType t, @Nonnull Point2D c, double rot,
             double xFactor, double yFactor, @Nonnull LayoutEditor layoutEditor, int v) {
         super(id, t, c, rot, xFactor, yFactor, layoutEditor, 1);
-        
+
         editor = new jmri.jmrit.display.layoutEditor.LayoutEditorDialogs.LayoutXOverEditor(layoutEditor);
     }
-    
+
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutXOver.class);
 }
