+++ java/src/jmri/jmrit/display/layoutEditor/LayoutXOverView.java
@@ -1,20 +1,12 @@
 package jmri.jmrit.display.layoutEditor;
 
-import java.awt.Color;
-import java.awt.Graphics2D;
-import java.awt.event.MouseEvent;
-import java.awt.geom.*;
-import java.util.*;
 import javax.annotation.*;
-import javax.swing.JPopupMenu;
-import jmri.*;
-import jmri.util.*;
 
 /**
  * MVC View component for the LayoutXOver class.
  *
  * @author Bob Jacobsen  Copyright (c) 2020
- * 
+ *
  */
 public class LayoutXOverView extends LayoutTurnoutView {
 
@@ -26,7 +18,7 @@ public class LayoutXOverView extends LayoutTurnoutView {
         super(xover);
         // this.xover = xover;
     }
-        
+
     // final private LayoutXOver xover;
 
     // private final static org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(LayoutXOverView.class);
