+++ java/src/jmri/jmrit/display/layoutEditor/TrackSegment.java
@@ -2,9 +2,7 @@ package jmri.jmrit.display.layoutEditor;
 
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 
-import java.awt.BasicStroke;
 import java.awt.Color;
-import java.awt.Graphics2D;
 import java.awt.event.ActionEvent;
 import java.awt.event.MouseEvent;
 import java.awt.geom.*;
@@ -2310,534 +2308,15 @@ public class TrackSegment extends LayoutTrack {
      * {@inheritDoc}
      */
     @Override
-    protected void draw1(Graphics2D g2, boolean isMain, boolean isBlock) {
-//   if (getName().equals("T15")) {
-//       log.debug("STOP");
-//   }
-        if (!isBlock && isDashed() && getLayoutBlock() != null) {
-            // Skip the dashed rail layer, the block layer will display the dashed track
-            // This removes random rail fragments from between the block dashes
-            return;
-        }
-        if (isMain == mainline) {
-            if (isBlock) {
-                setColorForTrackBlock(g2, getLayoutBlock());
-            }
-            if (isArc()) {
-                calculateTrackSegmentAngle();
-                g2.draw(new Arc2D.Double(getCX(), getCY(), getCW(), getCH(), getStartAdj(), getTmpAngle(), Arc2D.OPEN));
-                trackRedrawn();
-            } else if (isBezier()) {
-                Point2D[] points = getBezierPoints();
-                MathUtil.drawBezier(g2, points);
-            } else {
-                Point2D end1 = LayoutEditor.getCoords(getConnect1(), getType1());
-                Point2D end2 = LayoutEditor.getCoords(getConnect2(), getType2());
-
-                g2.draw(new Line2D.Double(end1, end2));
-            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void draw2(Graphics2D g2, boolean isMain, float railDisplacement) {
-//   if (getName().equals("T5")) {
-//       log.debug("STOP");
-//   }
-        if (isDashed() && getLayoutBlock() != null) {
-            // Skip the dashed rail layer, the block layer will display the dashed track
-            // This removes random rail fragments from between the block dashes
-            return;
-        }
-        if (isMain == mainline) {
-            if (isArc()) {
-                calculateTrackSegmentAngle();
-                Rectangle2D cRectangle2D = new Rectangle2D.Double(
-                        getCX(), getCY(), getCW(), getCH());
-                Rectangle2D tRectangle2D = MathUtil.inset(cRectangle2D, -railDisplacement);
-                double startAdj = getStartAdj(), tmpAngle = getTmpAngle();
-                g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                        tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                        startAdj, tmpAngle, Arc2D.OPEN));
-                tRectangle2D = MathUtil.inset(cRectangle2D, +railDisplacement);
-                g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                        tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                        startAdj, tmpAngle, Arc2D.OPEN));
-                trackRedrawn();
-            } else if (isBezier()) {
-                Point2D[] points = getBezierPoints();
-                MathUtil.drawBezier(g2, points, -railDisplacement);
-                MathUtil.drawBezier(g2, points, +railDisplacement);
-            } else {
-                Point2D end1 = LayoutEditor.getCoords(getConnect1(), getType1());
-                Point2D end2 = LayoutEditor.getCoords(getConnect2(), getType2());
-
-                Point2D delta = MathUtil.subtract(end2, end1);
-                Point2D vector = MathUtil.normalize(delta, railDisplacement);
-                vector = MathUtil.orthogonal(vector);
-
-                Point2D ep1L = MathUtil.add(end1, vector);
-                Point2D ep2L = MathUtil.add(end2, vector);
-                g2.draw(new Line2D.Double(ep1L, ep2L));
-
-                Point2D ep1R = MathUtil.subtract(end1, vector);
-                Point2D ep2R = MathUtil.subtract(end2, vector);
-                g2.draw(new Line2D.Double(ep1R, ep2R));
-            }
-        }
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void highlightUnconnected(Graphics2D g2, HitPointType selectedType) {
-        // TrackSegments are always connected
-        // nothing to see here... move along...
-    }
-
-    @Override
-    protected void drawEditControls(Graphics2D g2) {
-        g2.setColor(Color.black);
-        if (isShowConstructionLines()) {
-            Point2D ep1 = LayoutEditor.getCoords(getConnect1(), getType1());
-            Point2D ep2 = LayoutEditor.getCoords(getConnect2(), getType2());
-            if (isCircle()) {
-                // draw radiuses
-                Point2D circleCenterPoint = getCoordsCenterCircle();
-                g2.draw(new Line2D.Double(circleCenterPoint, ep1));
-                g2.draw(new Line2D.Double(circleCenterPoint, ep2));
-                // Draw a circle and square at the circles centre, that
-                // allows the user to change the angle by dragging the mouse.
-                g2.draw(trackEditControlCircleAt(circleCenterPoint));
-                g2.draw(layoutEditor.layoutEditorControlRectAt(circleCenterPoint));
-            } else if (isBezier()) {
-                // draw construction lines and control circles
-                Point2D lastPt = ep1;
-                for (Point2D bcp : bezierControlPoints) {
-                    g2.draw(new Line2D.Double(lastPt, bcp));
-                    lastPt = bcp;
-                    g2.draw(layoutEditor.layoutEditorControlRectAt(bcp));
-                }
-                g2.draw(new Line2D.Double(lastPt, ep2));
-            }
-        }
-        g2.draw(trackEditControlCircleAt(getCentreSeg()));
-    }   // drawEditControls
-
-    @Override
-    protected void drawTurnoutControls(Graphics2D g2) {
-        // TrackSegments don't have turnout controls...
-        // nothing to see here... move along...
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
     public void reCheckBlockBoundary() {
         // nothing to see here... move along...
     }
-
-    /**
-     * {@inheritDoc}
-     */
-    @Override
-    protected void drawDecorations(Graphics2D g2) {
-//   if (getName().equals("T5")) {
-//       log.debug("STOP");
-//   }
-
-// get end points and calculate start/stop angles (in radians)
-        Point2D ep1 = LayoutEditor.getCoords(getConnect1(), getType1());
-        Point2D ep2 = LayoutEditor.getCoords(getConnect2(), getType2());
-        Point2D p1, p2, p3, p4, p5, p6, p7;
-        Point2D p1P = ep1, p2P = ep2, p3P, p4P, p5P, p6P, p7P;
-        double startAngleRAD, stopAngleRAD;
-        if (isArc()) {
-            calculateTrackSegmentAngle();
-            double startAngleDEG = getStartAdj(), extentAngleDEG = getTmpAngle();
-            startAngleRAD = (Math.PI / 2.D) - Math.toRadians(startAngleDEG);
-            stopAngleRAD = (Math.PI / 2.D) - Math.toRadians(startAngleDEG + extentAngleDEG);
-            if (isFlip()) {
-                startAngleRAD += Math.PI;
-                stopAngleRAD += Math.PI;
-            } else {
-                double temp = startAngleRAD;
-                startAngleRAD = stopAngleRAD;
-                stopAngleRAD = temp;
-            }
-        } else if (isBezier()) {
-            Point2D cp0 = bezierControlPoints.get(0);
-            Point2D cpN = bezierControlPoints.get(bezierControlPoints.size() - 1);
-            startAngleRAD = (Math.PI / 2.D) - MathUtil.computeAngleRAD(cp0, ep1);
-            stopAngleRAD = (Math.PI / 2.D) - MathUtil.computeAngleRAD(ep2, cpN);
-        } else {
-            startAngleRAD = (Math.PI / 2.D) - MathUtil.computeAngleRAD(ep2, ep1);
-            stopAngleRAD = startAngleRAD;
-        }
-
-//
-// arrow decorations
-//
-        if (arrowStyle > 0) {
-            g2.setStroke(new BasicStroke(arrowLineWidth,
-                    BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER, 1.F));
-            g2.setColor(arrowColor);
-
-            // draw the start arrows
-            int offset = 1;
-            if (arrowEndStart) {
-                if (arrowDirIn) {
-                    offset = drawArrow(g2, ep1, Math.PI + startAngleRAD, false, offset);
-                }
-                if (arrowDirOut) {
-                    offset = drawArrow(g2, ep1, Math.PI + startAngleRAD, true, offset);
-                }
-            }
-
-            // draw the stop arrows
-            offset = 1;
-            if (arrowEndStop) {
-                if (arrowDirIn) {
-                    offset = drawArrow(g2, ep2, stopAngleRAD, false, offset);
-                }
-                if (arrowDirOut) {
-                    offset = drawArrow(g2, ep2, stopAngleRAD, true, offset);
-                }
-            }
-        }   // arrow decoration
-
-//
-// bridge decorations
-//
-        if (bridgeSideLeft || bridgeSideRight) {
-            float halfWidth = bridgeDeckWidth / 2.F;
-
-            g2.setStroke(new BasicStroke(bridgeLineWidth,
-                    BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER, 1.F));
-            g2.setColor(bridgeColor);
-
-            if (isArc()) {
-                calculateTrackSegmentAngle();
-                Rectangle2D cRectangle2D = new Rectangle2D.Double(
-                        getCX(), getCY(), getCW(), getCH());
-                double startAdj = getStartAdj(), tmpAngle = getTmpAngle();
-                if (bridgeSideLeft) {
-                    Rectangle2D tRectangle2D = MathUtil.inset(cRectangle2D, -halfWidth);
-                    g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                            tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                            startAdj, tmpAngle, Arc2D.OPEN));
-                }
-                if (bridgeSideRight) {
-                    Rectangle2D tRectangle2D = MathUtil.inset(cRectangle2D, +halfWidth);
-                    g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                            tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                            startAdj, tmpAngle, Arc2D.OPEN));
-                }
-            } else if (isBezier()) {
-                Point2D[] points = getBezierPoints();
-                if (bridgeSideLeft) {
-                    MathUtil.drawBezier(g2, points, -halfWidth);
-                }
-                if (bridgeSideRight) {
-                    MathUtil.drawBezier(g2, points, +halfWidth);
-                }
-            } else {
-                Point2D delta = MathUtil.subtract(ep2, ep1);
-                Point2D vector = MathUtil.normalize(delta, halfWidth);
-                vector = MathUtil.orthogonal(vector);
-
-                if (bridgeSideRight) {
-                    Point2D ep1R = MathUtil.add(ep1, vector);
-                    Point2D ep2R = MathUtil.add(ep2, vector);
-                    g2.draw(new Line2D.Double(ep1R, ep2R));
-                }
-
-                if (bridgeSideLeft) {
-                    Point2D ep1L = MathUtil.subtract(ep1, vector);
-                    Point2D ep2L = MathUtil.subtract(ep2, vector);
-                    g2.draw(new Line2D.Double(ep1L, ep2L));
-                }
-            }   // if isArc() {} else if isBezier() {} else...
-
-            if (isFlip()) {
-                boolean temp = bridgeSideRight;
-                bridgeSideRight = bridgeSideLeft;
-                bridgeSideLeft = temp;
-            }
-
-            if (bridgeHasEntry) {
-                if (bridgeSideRight) {
-                    p1 = new Point2D.Double(-bridgeApproachWidth, +bridgeApproachWidth + halfWidth);
-                    p2 = new Point2D.Double(0.0, +halfWidth);
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, startAngleRAD), ep1);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, startAngleRAD), ep1);
-                    g2.draw(new Line2D.Double(p1P, p2P));
-                }
-                if (bridgeSideLeft) {
-                    p1 = new Point2D.Double(-bridgeApproachWidth, -bridgeApproachWidth - halfWidth);
-                    p2 = new Point2D.Double(0.0, -halfWidth);
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, startAngleRAD), ep1);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, startAngleRAD), ep1);
-                    g2.draw(new Line2D.Double(p1P, p2P));
-                }
-            }
-            if (bridgeHasExit) {
-                if (bridgeSideRight) {
-                    p1 = new Point2D.Double(+bridgeApproachWidth, +bridgeApproachWidth + halfWidth);
-                    p2 = new Point2D.Double(0.0, +halfWidth);
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, stopAngleRAD), ep2);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, stopAngleRAD), ep2);
-                    g2.draw(new Line2D.Double(p1P, p2P));
-                }
-                if (bridgeSideLeft) {
-                    p1 = new Point2D.Double(+bridgeApproachWidth, -bridgeApproachWidth - halfWidth);
-                    p2 = new Point2D.Double(0.0, -halfWidth);
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, stopAngleRAD), ep2);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, stopAngleRAD), ep2);
-                    g2.draw(new Line2D.Double(p1P, p2P));
-                }
-            }
-
-            // if necessary flip these back
-            if (isFlip()) {
-                boolean temp = bridgeSideRight;
-                bridgeSideRight = bridgeSideLeft;
-                bridgeSideLeft = temp;
-            }
-        }
-
-//
-// end bumper decorations
-//
-        if (bumperEndStart || bumperEndStop) {
-            g2.setStroke(new BasicStroke(bumperLineWidth,
-                    BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER, 1.F));
-            g2.setColor(bumperColor);
-
-            float halfLength = bumperLength / 2.F;
-
-            if (bumperFlipped) {
-                double temp = startAngleRAD;
-                startAngleRAD = stopAngleRAD;
-                stopAngleRAD = temp;
-            }
-
-            // common points
-            p1 = new Point2D.Double(0.F, -halfLength);
-            p2 = new Point2D.Double(0.F, +halfLength);
-
-            if (bumperEndStart) {
-                p1P = MathUtil.add(MathUtil.rotateRAD(p1, startAngleRAD), ep1);
-                p2P = MathUtil.add(MathUtil.rotateRAD(p2, startAngleRAD), ep1);
-                // draw cross tie
-                g2.draw(new Line2D.Double(p1P, p2P));
-            }
-            if (bumperEndStop) {
-                p1P = MathUtil.add(MathUtil.rotateRAD(p1, stopAngleRAD), ep2);
-                p2P = MathUtil.add(MathUtil.rotateRAD(p2, stopAngleRAD), ep2);
-                // draw cross tie
-                g2.draw(new Line2D.Double(p1P, p2P));
-            }
-        }   // if (bumperEndStart || bumperEndStop)
-
-//
-// tunnel decorations
-//
-        if (tunnelSideRight || tunnelSideLeft) {
-            float halfWidth = tunnelFloorWidth / 2.F;
-            g2.setStroke(new BasicStroke(tunnelLineWidth,
-                    BasicStroke.CAP_BUTT, BasicStroke.JOIN_BEVEL, 10.F,
-                    new float[]{6.F, 4.F}, 0));
-            g2.setColor(tunnelColor);
-
-            if (isArc()) {
-                calculateTrackSegmentAngle();
-                Rectangle2D cRectangle2D = new Rectangle2D.Double(
-                        getCX(), getCY(), getCW(), getCH());
-                double startAngleDEG = getStartAdj(), extentAngleDEG = getTmpAngle();
-                if (tunnelSideRight) {
-                    Rectangle2D tRectangle2D = MathUtil.inset(cRectangle2D, +halfWidth);
-                    g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                            tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                            startAngleDEG, extentAngleDEG, Arc2D.OPEN));
-                }
-                if (tunnelSideLeft) {
-                    Rectangle2D tRectangle2D = MathUtil.inset(cRectangle2D, -halfWidth);
-                    g2.draw(new Arc2D.Double(tRectangle2D.getX(), tRectangle2D.getY(),
-                            tRectangle2D.getWidth(), tRectangle2D.getHeight(),
-                            startAngleDEG, extentAngleDEG, Arc2D.OPEN));
-                }
-                trackRedrawn();
-            } else if (isBezier()) {
-                Point2D[] points = getBezierPoints();
-                if (tunnelSideRight) {
-                    MathUtil.drawBezier(g2, points, +halfWidth);
-                }
-                if (tunnelSideLeft) {
-                    MathUtil.drawBezier(g2, points, -halfWidth);
-                }
-            } else {
-                Point2D delta = MathUtil.subtract(ep2, ep1);
-                Point2D vector = MathUtil.normalize(delta, halfWidth);
-                vector = MathUtil.orthogonal(vector);
-
-                if (tunnelSideRight) {
-                    Point2D ep1L = MathUtil.add(ep1, vector);
-                    Point2D ep2L = MathUtil.add(ep2, vector);
-                    g2.draw(new Line2D.Double(ep1L, ep2L));
-                }
-                if (tunnelSideLeft) {
-                    Point2D ep1R = MathUtil.subtract(ep1, vector);
-                    Point2D ep2R = MathUtil.subtract(ep2, vector);
-                    g2.draw(new Line2D.Double(ep1R, ep2R));
-                }
-            }   // if isArc() {} else if isBezier() {} else...
-
-            g2.setStroke(new BasicStroke(tunnelLineWidth,
-                    BasicStroke.CAP_BUTT, BasicStroke.JOIN_MITER, 1.F));
-            g2.setColor(tunnelColor);
-
-            // don't let tunnelEntranceWidth be less than tunnelFloorWidth + 6
-            tunnelEntranceWidth = Math.max(tunnelEntranceWidth, tunnelFloorWidth + 6);
-
-            double halfEntranceWidth = tunnelEntranceWidth / 2.0;
-            double halfFloorWidth = tunnelFloorWidth / 2.0;
-            double halfDiffWidth = halfEntranceWidth - halfFloorWidth;
-
-            if (isFlip()) {
-                boolean temp = tunnelSideRight;
-                tunnelSideRight = tunnelSideLeft;
-                tunnelSideLeft = temp;
-            }
-
-            if (tunnelHasEntry) {
-                if (tunnelSideRight) {
-                    p1 = new Point2D.Double(0.0, 0.0);
-                    p2 = new Point2D.Double(0.0, +halfFloorWidth);
-                    p3 = new Point2D.Double(0.0, +halfEntranceWidth);
-                    p4 = new Point2D.Double(-halfEntranceWidth - halfFloorWidth, +halfEntranceWidth);
-                    p5 = new Point2D.Double(-halfEntranceWidth - halfFloorWidth, +halfEntranceWidth - halfDiffWidth);
-                    p6 = new Point2D.Double(-halfFloorWidth, +halfEntranceWidth - halfDiffWidth);
-                    p7 = new Point2D.Double(-halfDiffWidth, 0.0);
-
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, startAngleRAD), ep1);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, startAngleRAD), ep1);
-                    p3P = MathUtil.add(MathUtil.rotateRAD(p3, startAngleRAD), ep1);
-                    p4P = MathUtil.add(MathUtil.rotateRAD(p4, startAngleRAD), ep1);
-                    p5P = MathUtil.add(MathUtil.rotateRAD(p5, startAngleRAD), ep1);
-                    p6P = MathUtil.add(MathUtil.rotateRAD(p6, startAngleRAD), ep1);
-                    p7P = MathUtil.add(MathUtil.rotateRAD(p7, startAngleRAD), ep1);
-
-                    GeneralPath path = new GeneralPath();
-                    path.moveTo(p1P.getX(), p1P.getY());
-                    path.lineTo(p2P.getX(), p2P.getY());
-                    path.quadTo(p3P.getX(), p3P.getY(), p4P.getX(), p4P.getY());
-                    path.lineTo(p5P.getX(), p5P.getY());
-                    path.quadTo(p6P.getX(), p6P.getY(), p7P.getX(), p7P.getY());
-                    path.closePath();
-                    g2.draw(path);
-                }
-                if (tunnelSideLeft) {
-                    p1 = new Point2D.Double(0.0, 0.0);
-                    p2 = new Point2D.Double(0.0, -halfFloorWidth);
-                    p3 = new Point2D.Double(0.0, -halfEntranceWidth);
-                    p4 = new Point2D.Double(-halfEntranceWidth - halfFloorWidth, -halfEntranceWidth);
-                    p5 = new Point2D.Double(-halfEntranceWidth - halfFloorWidth, -halfEntranceWidth + halfDiffWidth);
-                    p6 = new Point2D.Double(-halfFloorWidth, -halfEntranceWidth + halfDiffWidth);
-                    p7 = new Point2D.Double(-halfDiffWidth, 0.0);
-
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, startAngleRAD), ep1);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, startAngleRAD), ep1);
-                    p3P = MathUtil.add(MathUtil.rotateRAD(p3, startAngleRAD), ep1);
-                    p4P = MathUtil.add(MathUtil.rotateRAD(p4, startAngleRAD), ep1);
-                    p5P = MathUtil.add(MathUtil.rotateRAD(p5, startAngleRAD), ep1);
-                    p6P = MathUtil.add(MathUtil.rotateRAD(p6, startAngleRAD), ep1);
-                    p7P = MathUtil.add(MathUtil.rotateRAD(p7, startAngleRAD), ep1);
-
-                    GeneralPath path = new GeneralPath();
-                    path.moveTo(p1P.getX(), p1P.getY());
-                    path.lineTo(p2P.getX(), p2P.getY());
-                    path.quadTo(p3P.getX(), p3P.getY(), p4P.getX(), p4P.getY());
-                    path.lineTo(p5P.getX(), p5P.getY());
-                    path.quadTo(p6P.getX(), p6P.getY(), p7P.getX(), p7P.getY());
-                    path.closePath();
-                    g2.draw(path);
-                }
-            }
-            if (tunnelHasExit) {
-                if (tunnelSideRight) {
-                    p1 = new Point2D.Double(0.0, 0.0);
-                    p2 = new Point2D.Double(0.0, +halfFloorWidth);
-                    p3 = new Point2D.Double(0.0, +halfEntranceWidth);
-                    p4 = new Point2D.Double(halfEntranceWidth + halfFloorWidth, +halfEntranceWidth);
-                    p5 = new Point2D.Double(halfEntranceWidth + halfFloorWidth, +halfEntranceWidth - halfDiffWidth);
-                    p6 = new Point2D.Double(halfFloorWidth, +halfEntranceWidth - halfDiffWidth);
-                    p7 = new Point2D.Double(halfDiffWidth, 0.0);
-
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, stopAngleRAD), ep2);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, stopAngleRAD), ep2);
-                    p3P = MathUtil.add(MathUtil.rotateRAD(p3, stopAngleRAD), ep2);
-                    p4P = MathUtil.add(MathUtil.rotateRAD(p4, stopAngleRAD), ep2);
-                    p5P = MathUtil.add(MathUtil.rotateRAD(p5, stopAngleRAD), ep2);
-                    p6P = MathUtil.add(MathUtil.rotateRAD(p6, stopAngleRAD), ep2);
-                    p7P = MathUtil.add(MathUtil.rotateRAD(p7, stopAngleRAD), ep2);
-
-                    GeneralPath path = new GeneralPath();
-                    path.moveTo(p1P.getX(), p1P.getY());
-                    path.lineTo(p2P.getX(), p2P.getY());
-                    path.quadTo(p3P.getX(), p3P.getY(), p4P.getX(), p4P.getY());
-                    path.lineTo(p5P.getX(), p5P.getY());
-                    path.quadTo(p6P.getX(), p6P.getY(), p7P.getX(), p7P.getY());
-                    path.closePath();
-                    g2.draw(path);
-                }
-                if (tunnelSideLeft) {
-                    p1 = new Point2D.Double(0.0, 0.0);
-                    p2 = new Point2D.Double(0.0, -halfFloorWidth);
-                    p3 = new Point2D.Double(0.0, -halfEntranceWidth);
-                    p4 = new Point2D.Double(halfEntranceWidth + halfFloorWidth, -halfEntranceWidth);
-                    p5 = new Point2D.Double(halfEntranceWidth + halfFloorWidth, -halfEntranceWidth + halfDiffWidth);
-                    p6 = new Point2D.Double(halfFloorWidth, -halfEntranceWidth + halfDiffWidth);
-                    p7 = new Point2D.Double(halfDiffWidth, 0.0);
-
-                    p1P = MathUtil.add(MathUtil.rotateRAD(p1, stopAngleRAD), ep2);
-                    p2P = MathUtil.add(MathUtil.rotateRAD(p2, stopAngleRAD), ep2);
-                    p3P = MathUtil.add(MathUtil.rotateRAD(p3, stopAngleRAD), ep2);
-                    p4P = MathUtil.add(MathUtil.rotateRAD(p4, stopAngleRAD), ep2);
-                    p5P = MathUtil.add(MathUtil.rotateRAD(p5, stopAngleRAD), ep2);
-                    p6P = MathUtil.add(MathUtil.rotateRAD(p6, stopAngleRAD), ep2);
-                    p7P = MathUtil.add(MathUtil.rotateRAD(p7, stopAngleRAD), ep2);
-
-                    GeneralPath path = new GeneralPath();
-                    path.moveTo(p1P.getX(), p1P.getY());
-                    path.lineTo(p2P.getX(), p2P.getY());
-                    path.quadTo(p3P.getX(), p3P.getY(), p4P.getX(), p4P.getY());
-                    path.lineTo(p5P.getX(), p5P.getY());
-                    path.quadTo(p6P.getX(), p6P.getY(), p7P.getX(), p7P.getY());
-                    path.closePath();
-                    g2.draw(path);
-                }
-            }
-
-            // if necessary, put these back
-            if (isFlip()) {
-                boolean temp = tunnelSideRight;
-                tunnelSideRight = tunnelSideLeft;
-                tunnelSideLeft = temp;
-            }
-        }
-    }   // drawDecorations
 
     /*
     * getBezierPoints
     * @return the points to pass to MathUtil.drawBezier(...)
      */
-    private Point2D[] getBezierPoints() {
+    public Point2D[] getBezierPoints() {
         Point2D ep1 = LayoutEditor.getCoords(getConnect1(), getType1());
         Point2D ep2 = LayoutEditor.getCoords(getConnect2(), getType2());
         int cnt = bezierControlPoints.size() + 2;
@@ -2849,157 +2328,6 @@ public class TrackSegment extends LayoutTrack {
         points[cnt - 1] = ep2;
         return points;
     }
-
-    private int drawArrow(
-            Graphics2D g2,
-            Point2D ep,
-            double angleRAD,
-            boolean dirOut,
-            int offset) {
-        Point2D p1, p2, p3, p4, p5, p6;
-        switch (arrowStyle) {
-            default: {
-                arrowStyle = 0;
-                break;
-            }
-            case 0: {
-                break;
-            }
-            case 1: {
-                if (dirOut) {
-                    p1 = new Point2D.Double(offset, -arrowLength);
-                    p2 = new Point2D.Double(offset + arrowLength, 0.0);
-                    p3 = new Point2D.Double(offset, +arrowLength);
-                } else {
-                    p1 = new Point2D.Double(offset + arrowLength, -arrowLength);
-                    p2 = new Point2D.Double(offset, 0.0);
-                    p3 = new Point2D.Double(offset + arrowLength, +arrowLength);
-                }
-                p1 = MathUtil.add(MathUtil.rotateRAD(p1, angleRAD), ep);
-                p2 = MathUtil.add(MathUtil.rotateRAD(p2, angleRAD), ep);
-                p3 = MathUtil.add(MathUtil.rotateRAD(p3, angleRAD), ep);
-
-                g2.draw(new Line2D.Double(p1, p2));
-                g2.draw(new Line2D.Double(p2, p3));
-                offset += arrowLength + arrowGap;
-                break;
-            }
-            case 2: {
-                if (dirOut) {
-                    p1 = new Point2D.Double(offset, -arrowLength);
-                    p2 = new Point2D.Double(offset + arrowLength, 0.0);
-                    p3 = new Point2D.Double(offset, +arrowLength);
-                    p4 = new Point2D.Double(offset + arrowLineWidth + arrowGap, -arrowLength);
-                    p5 = new Point2D.Double(offset + arrowLineWidth + arrowGap + arrowLength, 0.0);
-                    p6 = new Point2D.Double(offset + arrowLineWidth + arrowGap, +arrowLength);
-                } else {
-                    p1 = new Point2D.Double(offset + arrowLength, -arrowLength);
-                    p2 = new Point2D.Double(offset, 0.0);
-                    p3 = new Point2D.Double(offset + arrowLength, +arrowLength);
-                    p4 = new Point2D.Double(offset + arrowLineWidth + arrowGap + arrowLength, -arrowLength);
-                    p5 = new Point2D.Double(offset + arrowLineWidth + arrowGap, 0.0);
-                    p6 = new Point2D.Double(offset + arrowLineWidth + arrowGap + arrowLength, +arrowLength);
-                }
-                p1 = MathUtil.add(MathUtil.rotateRAD(p1, angleRAD), ep);
-                p2 = MathUtil.add(MathUtil.rotateRAD(p2, angleRAD), ep);
-                p3 = MathUtil.add(MathUtil.rotateRAD(p3, angleRAD), ep);
-                p4 = MathUtil.add(MathUtil.rotateRAD(p4, angleRAD), ep);
-                p5 = MathUtil.add(MathUtil.rotateRAD(p5, angleRAD), ep);
-                p6 = MathUtil.add(MathUtil.rotateRAD(p6, angleRAD), ep);
-
-                g2.draw(new Line2D.Double(p1, p2));
-                g2.draw(new Line2D.Double(p2, p3));
-                g2.draw(new Line2D.Double(p4, p5));
-                g2.draw(new Line2D.Double(p5, p6));
-                offset += arrowLength + (2 * (arrowLineWidth + arrowGap));
-                break;
-            }
-            case 3: {
-                if (dirOut) {
-                    p1 = new Point2D.Double(offset, -arrowLength);
-                    p2 = new Point2D.Double(offset + arrowLength, 0.0);
-                    p3 = new Point2D.Double(offset, +arrowLength);
-                } else {
-                    p1 = new Point2D.Double(offset + arrowLength, -arrowLength);
-                    p2 = new Point2D.Double(offset, 0.0);
-                    p3 = new Point2D.Double(offset + arrowLength, +arrowLength);
-                }
-                p1 = MathUtil.add(MathUtil.rotateRAD(p1, angleRAD), ep);
-                p2 = MathUtil.add(MathUtil.rotateRAD(p2, angleRAD), ep);
-                p3 = MathUtil.add(MathUtil.rotateRAD(p3, angleRAD), ep);
-
-                GeneralPath path = new GeneralPath();
-                path.moveTo(p1.getX(), p1.getY());
-                path.lineTo(p2.getX(), p2.getY());
-                path.lineTo(p3.getX(), p3.getY());
-                path.closePath();
-                if (arrowLineWidth > 1) {
-                    g2.fill(path);
-                } else {
-                    g2.draw(path);
-                }
-                offset += arrowLength + arrowGap;
-                break;
-            }
-            case 4: {
-                if (dirOut) {
-                    p1 = new Point2D.Double(offset, 0.0);
-                    p2 = new Point2D.Double(offset + (2 * arrowLength), -arrowLength);
-                    p3 = new Point2D.Double(offset + (3 * arrowLength), 0.0);
-                    p4 = new Point2D.Double(offset + (2 * arrowLength), +arrowLength);
-                } else {
-                    p1 = new Point2D.Double(offset, 0.0);
-                    p2 = new Point2D.Double(offset + (4 * arrowLength), -arrowLength);
-                    p3 = new Point2D.Double(offset + (3 * arrowLength), 0.0);
-                    p4 = new Point2D.Double(offset + (4 * arrowLength), +arrowLength);
-                }
-                p1 = MathUtil.add(MathUtil.rotateRAD(p1, angleRAD), ep);
-                p2 = MathUtil.add(MathUtil.rotateRAD(p2, angleRAD), ep);
-                p3 = MathUtil.add(MathUtil.rotateRAD(p3, angleRAD), ep);
-                p4 = MathUtil.add(MathUtil.rotateRAD(p4, angleRAD), ep);
-
-                g2.draw(new Line2D.Double(p1, p3));
-                g2.draw(new Line2D.Double(p2, p3));
-                g2.draw(new Line2D.Double(p3, p4));
-
-                offset += (3 * arrowLength) + arrowGap;
-                break;
-            }
-            case 5: {
-                if (dirOut) {
-                    p1 = new Point2D.Double(offset, 0.0);
-                    p2 = new Point2D.Double(offset + (2 * arrowLength), -arrowLength);
-                    p3 = new Point2D.Double(offset + (3 * arrowLength), 0.0);
-                    p4 = new Point2D.Double(offset + (2 * arrowLength), +arrowLength);
-                } else {
-                    p1 = new Point2D.Double(offset, 0.0);
-                    p2 = new Point2D.Double(offset + (4 * arrowLength), -arrowLength);
-                    p3 = new Point2D.Double(offset + (3 * arrowLength), 0.0);
-                    p4 = new Point2D.Double(offset + (4 * arrowLength), +arrowLength);
-                }
-                p1 = MathUtil.add(MathUtil.rotateRAD(p1, angleRAD), ep);
-                p2 = MathUtil.add(MathUtil.rotateRAD(p2, angleRAD), ep);
-                p3 = MathUtil.add(MathUtil.rotateRAD(p3, angleRAD), ep);
-                p4 = MathUtil.add(MathUtil.rotateRAD(p4, angleRAD), ep);
-
-                GeneralPath path = new GeneralPath();
-                path.moveTo(p4.getX(), p4.getY());
-                path.lineTo(p2.getX(), p2.getY());
-                path.lineTo(p3.getX(), p3.getY());
-                path.closePath();
-                if (arrowLineWidth > 1) {
-                    g2.fill(path);
-                } else {
-                    g2.draw(path);
-                }
-                g2.draw(new Line2D.Double(p1, p3));
-
-                offset += (3 * arrowLength) + arrowGap;
-                break;
-            }
-        }
-        return offset;
-    }   // drawArrow
 
     /*======================*\
     |* decoration accessors *|
