+++ java/test/jmri/jmrit/display/layoutEditor/LayoutTrackViewTest.java
@@ -1,5 +1,9 @@
 package jmri.jmrit.display.layoutEditor;
 
+import java.awt.GraphicsEnvironment;
+
+import jmri.util.JUnitUtil;
+
 import org.junit.jupiter.api.*;
 
 /**
@@ -8,10 +12,5 @@ import org.junit.jupiter.api.*;
  * @author Bob Jacobsen Copyright (C) 2016
  */
 public class LayoutTrackViewTest {
-
-    @Test
-    public void testCtor() {
-        new LayoutTrackView(null);
-    }
 
 }
